api/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0

# Database
sqlalchemy==2.0.29
aioodbc==0.5.0
pyodbc==5.1.0
alembic==1.13.1

# Dev/Test
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
